requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
iniconfig==2.0.0
kombu==5.3.1
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
pytest==7.4.0
pytest-celery==0.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
